// old/packages/waas/src/clients/intent.gen.ts
export enum IntentName {
  initiateAuth = 'initiateAuth',
  openSession = 'openSession',
  closeSession = 'closeSession',
  validateSession = 'validateSession',
  finishValidateSession = 'finishValidateSession',
  listSessions = 'listSessions',
  getSession = 'getSession',
  sessionAuthProof = 'sessionAuthProof',
  feeOptions = 'feeOptions',
  signMessage = 'signMessage',
  signTypedData = 'signTypedData',
  sendTransaction = 'sendTransaction',
  getTransactionReceipt = 'getTransactionReceipt',
  federateAccount = 'federateAccount',
  removeAccount = 'removeAccount',
  listAccounts = 'listAccounts',
  getIdToken = 'getIdToken',
  adoptChildWallet = 'adoptChildWallet',
  getAdopter = 'getAdopter',
  confirmIntent = 'confirmIntent',
  getConfirmationStatus = 'getConfirmationStatus'
}
// ...
export enum FeeTokenType {
  unknown = 'unknown',
  erc20Token = 'erc20Token',
  erc1155Token = 'erc1155Token'
}
// ...
export interface IntentDataFeeOptions {
  network: string
  wallet: string
  identifier: string
  transactions: Array<any>
}

export interface IntentDataSendTransaction {
  network: string
  wallet: string
  identifier: string
  transactions: Array<any>
  transactionsFeeQuote?: string
}

export interface IntentDataGetTransactionReceipt {
  network: string
  wallet: string
  metaTxHash: string
}
// ...
export interface TransactionRaw {
  type: string
  to: string
  value?: string
  data: string
}

export interface AbiData {
  abi: string
  func?: string
  args: Array<any>
}

export interface TransactionERC20 {
  type: string
  tokenAddress: string
  to: string
  value: string
}

export interface TransactionERC721 {
  type: string
  tokenAddress: string
  to: string
  id: string
  safe?: boolean
  data?: string
}

export interface TransactionERC1155Value {
  id: string
  amount: string
}

export interface TransactionDelayedEncode {
  type: string
  to: string
  value: string
  data: any
}

export interface TransactionContractCall {
  type: string
  to: string
  value?: string
  data: AbiData
}

export interface TransactionERC1155 {
  type: string
  tokenAddress: string
  to: string
  vals: Array<TransactionERC1155Value>
  data?: string
}
// ...
export interface FeeOption {
  token: FeeToken
  to: string
  value: string
  gasLimit: number
}

export interface FeeToken {
  chainId: number
  name: string
  symbol: string
  type: FeeTokenType
  decimals?: number
  logoURL: string
  contractAddress?: string
  tokenID?: string
}